Fold BaseOAuthValidator into GenericOAuthValidator
