+++ src/Tgstation.Server.Host/Security/OAuth/GenericOAuthValidator.cs
@@ -2,12 +2,14 @@
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Net.Http.Headers;
+using System.Net.Mime;
 using System.Threading;
 using System.Threading.Tasks;
 
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using Newtonsoft.Json.Serialization;
 
 using Tgstation.Server.Api;
 using Tgstation.Server.Api.Models;
@@ -19,8 +21,21 @@ namespace Tgstation.Server.Host.Security.OAuth
 	/// <summary>
 	/// <see cref="IOAuthValidator"/> for generic OAuth2 endpoints.
 	/// </summary>
-	abstract class GenericOAuthValidator : BaseOAuthValidator
+	abstract class GenericOAuthValidator : IOAuthValidator
 	{
+		/// <inheritdoc />
+		public abstract OAuthProvider Provider { get; }
+
+		/// <summary>
+		/// The <see cref="ILogger"/> for the <see cref="GenericOAuthValidator"/>.
+		/// </summary>
+		protected ILogger<GenericOAuthValidator> Logger { get; }
+
+		/// <summary>
+		/// The <see cref="OAuthConfiguration"/> for the <see cref="GenericOAuthValidator"/>.
+		/// </summary>
+		protected OAuthConfiguration OAuthConfiguration { get; }
+
 		/// <summary>
 		/// <see cref="Uri"/> to <see cref="HttpMethod.Post"/> to to get the access token.
 		/// </summary>
@@ -32,27 +47,48 @@ namespace Tgstation.Server.Host.Security.OAuth
 		protected abstract Uri UserInformationUrl { get; }
 
 		/// <summary>
+		/// The <see cref="IHttpClientFactory"/> for the <see cref="GenericOAuthValidator"/>.
+		/// </summary>
+		readonly IHttpClientFactory httpClientFactory;
+
+		/// <summary>
+		/// The <see cref="IAssemblyInformationProvider"/> for the <see cref="GenericOAuthValidator"/>.
+		/// </summary>
+		readonly IAssemblyInformationProvider assemblyInformationProvider;
+
+		/// <summary>
+		/// Gets <see cref="JsonSerializerSettings"/> that should be used.
+		/// </summary>
+		/// <returns>A new <see cref="JsonSerializerSettings"/> <see cref="object"/>.</returns>
+		protected static JsonSerializerSettings SerializerSettings() => new JsonSerializerSettings
+		{
+			ContractResolver = new DefaultContractResolver
+			{
+				NamingStrategy = new SnakeCaseNamingStrategy(),
+			},
+		};
+
+		/// <summary>
 		/// Initializes a new instance of the <see cref="GenericOAuthValidator"/> class.
 		/// </summary>
-		/// <param name="httpClientFactory">The <see cref="IHttpClientFactory"/> for the <see cref="BaseOAuthValidator"/>.</param>
-		/// <param name="assemblyInformationProvider">The <see cref="IAssemblyInformationProvider"/> for the <see cref="BaseOAuthValidator"/>.</param>
-		/// <param name="logger">The <see cref="ILogger"/> for the <see cref="BaseOAuthValidator"/>.</param>
-		/// <param name="oAuthConfiguration">The <see cref="OAuthConfiguration"/> for the <see cref="BaseOAuthValidator"/>.</param>
+		/// <param name="httpClientFactory">The value of <see cref="httpClientFactory"/>.</param>
+		/// <param name="assemblyInformationProvider">The value of <see cref="assemblyInformationProvider"/>.</param>
+		/// <param name="logger">The value of <see cref="Logger"/>.</param>
+		/// <param name="oAuthConfiguration">The value of <see cref="OAuthConfiguration"/>.</param>
 		public GenericOAuthValidator(
 			IHttpClientFactory httpClientFactory,
 			IAssemblyInformationProvider assemblyInformationProvider,
 			ILogger<GenericOAuthValidator> logger,
 			OAuthConfiguration oAuthConfiguration)
-			: base(
-				 httpClientFactory,
-				 assemblyInformationProvider,
-				 logger,
-				 oAuthConfiguration)
 		{
+			this.httpClientFactory = httpClientFactory ?? throw new ArgumentNullException(nameof(httpClientFactory));
+			this.assemblyInformationProvider = assemblyInformationProvider ?? throw new ArgumentNullException(nameof(assemblyInformationProvider));
+			Logger = logger ?? throw new ArgumentNullException(nameof(logger));
+			OAuthConfiguration = oAuthConfiguration ?? throw new ArgumentNullException(nameof(oAuthConfiguration));
 		}
 
 		/// <inheritdoc />
-		public override async Task<string> ValidateResponseCode(string code, CancellationToken cancellationToken)
+		public async Task<string> ValidateResponseCode(string code, CancellationToken cancellationToken)
 		{
 			using var httpClient = CreateHttpClient();
 			string tokenResponsePayload = null;
@@ -110,7 +146,7 @@ namespace Tgstation.Server.Host.Security.OAuth
 		}
 
 		/// <inheritdoc />
-		public override Task<OAuthProviderInfo> GetProviderInfo(CancellationToken cancellationToken) => Task.FromResult(
+		public Task<OAuthProviderInfo> GetProviderInfo(CancellationToken cancellationToken) => Task.FromResult(
 			new OAuthProviderInfo
 			{
 				ClientId = OAuthConfiguration.ClientId,
@@ -138,5 +174,25 @@ namespace Tgstation.Server.Host.Security.OAuth
 		/// <param name="code">The OAuth code from the browser.</param>
 		/// <returns>The <see cref="OAuthTokenRequest"/> to send to <see cref="TokenUrl"/>.</returns>
 		protected abstract OAuthTokenRequest CreateTokenRequest(string code);
+
+		/// <summary>
+		/// Create a new configured <see cref="HttpClient"/>.
+		/// </summary>
+		/// <returns>A new configured <see cref="HttpClient"/>.</returns>
+		HttpClient CreateHttpClient()
+		{
+			var httpClient = httpClientFactory.CreateClient();
+			try
+			{
+				httpClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue(MediaTypeNames.Application.Json));
+				httpClient.DefaultRequestHeaders.UserAgent.Add(assemblyInformationProvider.ProductInfoHeaderValue);
+				return httpClient;
+			}
+			catch
+			{
+				httpClient.Dispose();
+				throw;
+			}
+		}
 	}
 }
