+++ src/Tgstation.Server.Host/Security/OAuth/TGForumsOAuthValidator.cs
@@ -26,10 +26,10 @@ namespace Tgstation.Server.Host.Security.OAuth
 		/// <summary>
 		/// Initializes a new instance of the <see cref="TGForumsOAuthValidator"/> class.
 		/// </summary>
-		/// <param name="httpClientFactory">The <see cref="IHttpClientFactory"/> for the <see cref="BaseOAuthValidator"/>.</param>
-		/// <param name="assemblyInformationProvider">The <see cref="IAssemblyInformationProvider"/> for the <see cref="BaseOAuthValidator"/>.</param>
-		/// <param name="logger">The <see cref="ILogger"/> for the <see cref="BaseOAuthValidator"/>.</param>
-		/// <param name="oAuthConfiguration">The <see cref="OAuthConfiguration"/> for the <see cref="BaseOAuthValidator"/>.</param>
+		/// <param name="httpClientFactory">The <see cref="IHttpClientFactory"/> for the <see cref="GenericOAuthValidator"/>.</param>
+		/// <param name="assemblyInformationProvider">The <see cref="IAssemblyInformationProvider"/> for the <see cref="GenericOAuthValidator"/>.</param>
+		/// <param name="logger">The <see cref="ILogger"/> for the <see cref="GenericOAuthValidator"/>.</param>
+		/// <param name="oAuthConfiguration">The <see cref="OAuthConfiguration"/> for the <see cref="GenericOAuthValidator"/>.</param>
 		public TGForumsOAuthValidator(
 			IHttpClientFactory httpClientFactory,
 			IAssemblyInformationProvider assemblyInformationProvider,
